refactor(container): mv AvatarAdder to tool scope
