+++ src/components/ContributorList/index.js
@@ -10,7 +10,7 @@ import T from 'prop-types'
 import { buildLog } from '@/utils'
 import { withGuardian } from '@/hoc'
 
-import AvatarAdder from '@/containers/AvatarAdder'
+import AvatarAdder from '@/containers/tool/AvatarAdder'
 import Tooltip from '@/components/Tooltip'
 import GithubUserCard from '@/components/GithubUserCard'
