+++ src/containers/tool/AvatarAdder/AdderPanel.js
