+++ src/containers/tool/AvatarAdder/schema.js
