+++ src/containers/tool/AvatarAdder/styles/adder_panel.js
