+++ src/containers/tool/AvatarAdder/styles/index.js
