+++ src/containers/tool/AvatarAdder/tests/index.test.js
