+++ src/containers/tool/AvatarAdder/tests/store.test.js
